Progress save: further progress with BSA, BSA now installs/uninstalls (tested), fixed a major bug in settings/settingsnode with parsing values from the database, modified layout, added ability to embed null variables using semi-colon, CSRF finished, added ability to set flags in data.

+++ Code/App_Code/CMS/Plugins/Basic Site Auth/AccountEvent.cs
@@ -164,7 +164,7 @@ namespace CMS.BasicSiteAuth
             }
             else
             {
-                eventid = c.executeInsert(conn, "bsa_account_events", "eventid")[0].get2<int>("eventid");
+                eventid = (int)c.executeInsert(conn, "bsa_account_events", "eventid")[0].get2<long>("eventid");
                 persisted = true;
             }
             modified = false;
